Fixed bug resulting from 8-bit floats in AIA test data

+++ sunpy/image/tests/test_coalignment.py
@@ -34,7 +34,7 @@ def aia171_test_shift():
 
 @pytest.fixture
 def aia171_test_map_layer(aia171_test_map):
-    return aia171_test_map.data
+    return aia171_test_map.data.astype('float32')  # SciPy 1.4 requires at least 16-bit floats
 
 
 @pytest.fixture
